Move getAfterAuthPath()
